refactor: move tests for filters to higher level packages dir
